issue #2870: fix some warning for linux and mac

+++ cocos/2d/CCActionCatmullRom.cpp
@@ -43,7 +43,7 @@ NS_CC_BEGIN;
  *  Implementation of PointArray
  */
 
-PointArray* PointArray::create(unsigned int capacity)
+PointArray* PointArray::create(long capacity)
 {
     PointArray* ret = new PointArray();
     if (ret)
@@ -63,7 +63,7 @@ PointArray* PointArray::create(unsigned int capacity)
 }
 
 
-bool PointArray::initWithCapacity(unsigned int capacity)
+bool PointArray::initWithCapacity(long capacity)
 {
     _controlPoints = new vector<Point*>();
     
@@ -126,19 +126,19 @@ void PointArray::addControlPoint(Point controlPoint)
     _controlPoints->push_back(new Point(controlPoint.x, controlPoint.y));
 }
 
-void PointArray::insertControlPoint(Point &controlPoint, unsigned int index)
+void PointArray::insertControlPoint(Point &controlPoint, long index)
 {
     Point *temp = new Point(controlPoint.x, controlPoint.y);
     _controlPoints->insert(_controlPoints->begin() + index, temp);
 }
 
-Point PointArray::getControlPointAtIndex(unsigned int index)
+Point PointArray::getControlPointAtIndex(long index)
 {
     index = MIN(_controlPoints->size()-1, MAX(index, 0));
     return *(_controlPoints->at(index));
 }
 
-void PointArray::replaceControlPoint(cocos2d::Point &controlPoint, unsigned int index)
+void PointArray::replaceControlPoint(cocos2d::Point &controlPoint, long index)
 {
 
     Point *temp = _controlPoints->at(index);
@@ -146,7 +146,7 @@ void PointArray::replaceControlPoint(cocos2d::Point &controlPoint, unsigned int 
     temp->y = controlPoint.y;
 }
 
-void PointArray::removeControlPointAtIndex(unsigned int index)
+void PointArray::removeControlPointAtIndex(long index)
 {
     vector<Point*>::iterator iter = _controlPoints->begin() + index;
     Point* pRemovedPoint = *iter;
@@ -154,7 +154,7 @@ void PointArray::removeControlPointAtIndex(unsigned int index)
     delete pRemovedPoint;
 }
 
-unsigned int PointArray::count() const
+long PointArray::count() const
 {
     return _controlPoints->size();
 }
@@ -291,7 +291,7 @@ CardinalSplineTo* CardinalSplineTo::clone() const
 
 void CardinalSplineTo::update(float time)
 {
-    unsigned int p;
+    long p;
     float lt;
 	
 	// eg.
