+++ cocos/2d/CCActionCatmullRom.h
@@ -61,7 +61,7 @@ public:
     /** creates and initializes a Points array with capacity 
      * @js NA
      */
-    static PointArray* create(unsigned int capacity);
+    static PointArray* create(long capacity);
     
     /**
      * @js NA
@@ -77,7 +77,7 @@ public:
     /** initializes a Catmull Rom config with a capacity hint 
      * @js NA
      */
-    bool initWithCapacity(unsigned int capacity);
+    bool initWithCapacity(long capacity);
     
     /** appends a control point 
      * @js NA
@@ -87,27 +87,27 @@ public:
     /** inserts a controlPoint at index 
      * @js NA
      */
-    void insertControlPoint(Point &controlPoint, unsigned int index);
+    void insertControlPoint(Point &controlPoint, long index);
     
     /** replaces an existing controlPoint at index 
      * @js NA
      */
-    void replaceControlPoint(Point &controlPoint, unsigned int index);
+    void replaceControlPoint(Point &controlPoint, long index);
     
     /** get the value of a controlPoint at a given index 
      * @js NA
      */
-    Point getControlPointAtIndex(unsigned int index);
+    Point getControlPointAtIndex(long index);
     
     /** deletes a control point at a given index 
      * @js NA
      */
-    void removeControlPointAtIndex(unsigned int index);
+    void removeControlPointAtIndex(long index);
     
     /** returns the number of objects of the control point array 
      * @js NA
      */
-    unsigned int count() const;
+    long count() const;
     
     /** returns a new copy of the array reversed. User is responsible for releasing this copy 
      * @js NA
